Fix unevaluation diverging without burning fuel
By burning fuel when unevaluating projections, unevaluation loops such
as when unevaluating `foldNat A id (Succ {0}) <== A` are caught early.
This bug only showed up when the informativeness restriction was
removed, so no it is possible to remove the restriction and allow e.g.
list_index to be defined without a sketch.