Allow partial functions
Partial functions are very naturally permitted using holes. If no
constraints are left, the remaining holes can be seen as unreachable
code. This is useful when preconditions are taken into account in
example inputs, such that the partial behaviour of the target function
is unreachable. Even if we do not want to synthesize partial functions,
some total functions such as scanr use partial helper functions, which
we may need to synthesize. In other words, some functions contain
subproblems with preconditions, which may lead to unreachable code.